Consolidate UI building
