+++ exts/siborg.create.human/siborg/create/human/human_ui.py
@@ -221,87 +221,4 @@ class ParamPanel(ui.Frame):
         for model in self.models:
             model.destroy()
 
-class ButtonPanel:
-    """UI Widget that includes a list of assets currently applied to the human and
-    buttons for applying changes to the human in the stage
-    Attributes
-    ----------
-    mh_call : MHCaller
-        Wrapper object around Makehuman functions
-    param_panel : ParamPanel
-        Reference to UI list widget for resetting UI when human is reset
-    """
 
-    def __init__(self, mhcaller: MHCaller, toggle : ui.SimpleBoolModel, param_panel : ParamPanel, **kwargs):
-        """Constructs an instance of ButtonPanel, which contains a DropList for displaying currently applied assets, as well as the following buttons:
-        + Update in Scene - Updates the current human
-        + New Human - Abandons reference to previous human and creates a new one
-
-        Parameters
-        ----------
-        mhcaller : MHCaller
-            Wrapper object around Makehuman functions
-        toggle : ui.SimpleBoolModel
-            Model to toggle whether human should update immediately
-        param_panel : ParamPanel
-            Reference to UI list widget for resetting UI when human is reset
-        """
-        # Include instance of Makehuman wrapper class
-        self.mh_call = mhcaller
-        # Model to store whether changes should happen immediately
-        self.toggle = toggle
-        # Reference to UI list widget for resetting UI when human is reset
-        self.param_panel = param_panel
-
-        # Pass **kwargs to buildwidget so we can apply styling as though ButtonPanel
-        # extended a base ui class
-        self._build_widget(**kwargs)
-
-    def _build_widget(self, **kwargs):
-        """Build UI widget
-        """
-        with ui.VStack(**kwargs):
-            # Widget to list applied proxies TODO change to "Currently Applied Assets"
-            self.drop = DropList("Currently Applied Proxies", self.mh_call)
-            with ui.HStack(height=0):
-                # Toggle whether changes should propagate instantly
-                ui.Label("Update Instantly")
-                ui.CheckBox(self.toggle)
-            # Creates a new human in scene and resets modifiers and assets
-            ui.Button(
-                "New Human",
-                height=50,
-                clicked_fn=lambda: self.new_human(),
-            )
-            # Updates current human in omniverse scene
-            ui.Button(
-                "Update Meshes in Scene",
-                height=50,
-                clicked_fn=lambda: mh_usd.add_to_scene(self.mh_call),
-            )
-            # Apply skeleton
-            ui.Button(
-                "Bake and Rig",
-                height=50,
-                clicked_fn=lambda: mh_usd.add_to_scene(self.mh_call, True),
-            )
-
-    def new_human(self):
-        """Creates a new human in the stage. Makes calls to the Makehuman function wrapper MHCaller for resetting the human parameters and assets as well as flagging the human for renaming. Then creates a new human in the stage with the reset data. 
-        """
-        # Reset the human object in the makehuman wrapper. Also flags the human for
-        # name change to avoid overwriting existing humans
-        self.mh_call.reset_human()
-        # Reset list of applied assets
-        self.drop.model.update()
-        # Reset list of modifiers
-        self.param_panel.reset()
-        # Add the new, now reset human to the scene
-        mh_usd.add_to_scene(self.mh_call)
-
-    def destroy(self):
-        """Provides destructor for ButtonPanel instance. Currently doesn't do anything but avoid an exception when destroying all UI elements
-        """
-        # No superclass to destroy, no models to destroy
-        # TODO do we need to reference the DropList model for destruction?
-        pass
